requirements.txt
azure-core==1.30.1
azure-functions==1.17.0
azure-storage-blob==12.19.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.6
isodate==0.6.1
oauthlib==3.2.2
pycparser==2.22
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.10.0
urllib3==2.1.0
